Refactor: BackupService generateBackupFile 메소드 내부 구현 변경
기존의 Backup service 내부에 존재하던 구현내용을 CsvProvider 객체에 캡슐화했습니다.
